requirements.txt
colorzero==2.0
gpiozero==2.0.1
lgpio==0.2.2.0
numpy==2.2.4
pillow==11.1.0
smbus==1.1.post2
spidev==3.6
